refactor: move header_frame and sidebar_frame to frames directory
